Cleanup repository, set identifiers.py as old (now in symbolTable.py)
