+++ .gitignore
@@ -1,7 +1,6 @@
 
 
 __pycache__/
-generatedFiles/
 !generatedFiles/LLMcache.json
 tree-sitter-java/libtree-sitter-java.so
 .env
