--- file_data.csv
@@ -1,15 +1,0 @@
-filename,domains,subdomains
-src/main/java/org/jabref/logic/layout/format/HTMLChars.java,"Software Development and IT Operations, Data Structure, Utility","Software Development and IT Operations-Automated Testing, Data Structure-Linear Structures"
-src/main/java/org/jabref/logic/importer/fileformat/EndnoteXmlImporter.java,"Software Development and IT Operations, Databases, Logging, Input-Output, Data Structure, Parser, Error Handling, Utility",Software Development and IT Operations-Automated Testing
-src/main/java/org/jabref/logic/importer/util/MetaDataParser.java,"Software Development and IT Operations, Logging, Data Structure, Input-Output, Software Development and IT Operationsorg.apache.http.client.utils.URIBuilder, Parser, Error Handling, Utility","Software Development and IT Operations-Automated Testing, Software Development and IT Operations-Version Control, Data Structure-Linear Structures"
-src/main/java/org/jabref/model/entry/AuthorList.java,"Software Development and IT Operations, Data Structure, Parser, Utility","Data Structure-Linear Structures, Software Development and IT Operations-Automated Testing, Software Development and IT Operations-Version Control, Parser-Syntax Parsing"
-src/main/java/org/jabref/gui/StateManager.java,"Computer Graphics, Databases, Search, Data Structure, Utility",None Used
-src/main/java/org/jabref/gui/entryeditor/FieldsEditorTab.java,"Software Development and IT Operations, Data Structure, User Interface, Computer Graphics, Databases, Utility",None Used
-src/main/java/org/jabref/gui/fieldeditors/LinkedEntriesEditor.java,"Software Development and IT Operations, Data Structure, Databases, Computer Graphics, User Interface","Software Development and IT Operations-Configuration Management, Software Development and IT Operations-Version Control"
-src/main/java/org/jabref/model/entry/BibEntry.java,"Software Development and IT Operations, Data Structure, Logging, Databases, Utility, Event Handling, Language, Computer Graphics, Error Handling","Software Development and IT Operations-Automated Testing, Software Development and IT Operations-Version Control, Utility-Data Conversion, Data Structure-Data Manipulation, Utility-System Tools"
-src/main/java/org/jabref/model/database/BibDatabaseContext.java,"Data Structure, Databases, Event Handling, Input-Output, Software Development and IT Operations, Utility",Databases-Query Execution
-src/main/java/org/jabref/gui/fieldeditors/KeywordsEditor.java,"Software Development and IT Operations, Data Structure, User Interface",None Used
-src/main/java/org/jabref/gui/entryeditor/EntryEditorPreferences.java,"Software Development and IT Operations, Data Structure, Utility",None Used
-src/main/java/org/jabref/gui/externalfiles/ImportHandler.java,"Software Development and IT Operationsorg.apache.http.client.utils.URIBuilder, Software Development and IT Operations, Input-Output, Data Structure, Databases, User Interface, Utility",Input-Output-File Management
-src/main/java/org/jabref/gui/groups/GroupTreeViewModel.java,"Software Development and IT Operations, Computer Graphics, Utility, Data Structure, User Interface, Databases, Language","Software Development and IT Operations-Configuration Management, Software Development and IT Operations-Automated Testing"
-src/main/java/org/jabref/gui/groups/GroupNodeViewModel.java,"Software Development and IT Operations, Databases, Computer Graphics, Data Structure, Utility, Event Handling, Internationalization, User Interface","Software Development and IT Operations-Automated Testing, Software Development and IT Operations-Configuration Management"
