+++ generatedFiles/api_storage.csv
@@ -1,0 +1,19 @@
+key,label
+String-length,Software Development and IT Operations-Automated Testing
+String-substring,Software Development and IT Operations-Automated Testing
+StringBuilder-length,Data Structure-Linear Structures
+StringBuilder-append,Data Structure-Linear Structures
+StringBuilder-toString,Data Structure-Linear Structures
+BasePanel-cleanUp,User Interface-Interaction Design
+BasePanel-isPresent,User Interface-Interaction Design
+BasePanel-updateTableFont,User Interface-Layout Design
+BasePanel-getDatabase,User Interface-Interaction Design
+BasePanel-isModified,User Interface-Interaction Design
+BasePanel-getTabTitle,User Interface-Interaction Design
+BasePanel-getBibDatabaseContext,User Interface-Interaction Design
+BasePanel-getMainTable,User Interface-Interaction Design
+BasePanel-getSelectedEntries,User Interface-Interaction Design
+BasePanel-getUndoManager,User Interface-Interaction Design
+BasePanel-getCurrentSearchQuery,User Interface-Interaction Design
+BasePanel-updateSearchManager,User Interface-Interaction Design
+WaitForSaveFinishedDialog-showAndWait,Software Development and IT Operations-Configuration Management
