--- issues_data2 test.csv
@@ -1,130 +1,0 @@
-Repository,Issue Number,Open Date,Closed Date,Issue Author,Issue Title,Issue Body,Comment Bodies,Comment Authors,Number of Comments,Labels,PR Number,PR Merge Date,PR Author,PR Title,PR Body,PR Files
-repository0,11281,2024-05-06T18:24:02Z,2024-05-08T20:44:30Z,Siedlerchr,Position (width) of Preview inside entry editor not remembered,"### JabRef version
-
-Latest development branch build (please note build date below) 
-
-Issue was mentioned in the forum thread https://discourse.jabref.org/t/inability-of-jabref-to-restore-preferences-after-a-version-update/4296/3
-
-### Operating system
-
-macOS
-
-### Details on version and operating system
-
-_No response_
-
-### Checked with the latest development build (copy version output from About dialog)
-
-- [X] I made a backup of my libraries before testing the latest development version.
-- [X] I have tested the latest development version and the problem persists
-
-### Steps to reproduce the behaviour
-
-1. Open the entry editor
-2. Adjust the preview window position on the right, e.g. reduce the width 
-3. Close JabRef and restart
-=>  The window position is reset
-
-Step 2:
-<img width=""1246"" alt=""grafik"" src=""https://github.com/JabRef/jabref/assets/320228/fedc2497-5ffa-49e6-9835-55a598cb144c"">
-
-After restarting JabRef
-<img width=""1231"" alt=""grafik"" src=""https://github.com/JabRef/jabref/assets/320228/f91f23b4-6db1-4bcc-8d0e-ad38091bd1d1"">
-
-Implementation details:
-The EntryEditorPreferences need to be enhanced with a new Property for the Preview Position, similar to how it's done for the divider position which remembers the height. 
-
-
-### Appendix
-
-...
-<details>
-
-<summary>Log File</summary>
-
-```
-Paste an excerpt of your log file here
-```
-
-</details>
-",,,0,"bug, ui, entry-editor, good first issue",11285,2024-05-08T20:44:29Z,Siedlerchr,Store preview divider pos in entry editor,"Fixes https://github.com/JabRef/jabref/issues/11281
-
-<!-- 
-Describe the changes you have made here: what, why, ... 
-Link the issue that will be closed, e.g., ""Closes #333"".
-If your PR closes a koppor issue, link it using its URL, e.g., ""Closes https://github.com/koppor/jabref/issues/47"".
-""Closes"" is a keyword GitHub uses to link PRs with issues; do not change it.
-Don't reference an issue in the PR title because GitHub does not support auto-linking there.
--->
-
-### Mandatory checks
-
-<!-- 
-- Go through the list below. Please don't remove any items.
-- [x] done; [ ] not done / not applicable
--->
-
-- [x] Change in `CHANGELOG.md` described in a way that is understandable for the average user (if applicable)
-- [ ] Tests created for changes (if applicable)
-- [x] Manually tested changed features in running JabRef (always required)
-- [ ] Screenshots added in PR description (for UI changes)
-- [ ] [Checked developer's documentation](https://devdocs.jabref.org/): Is the information available and up to date? If not, I outlined it in this pull request.
-- [ ] [Checked documentation](https://docs.jabref.org/): Is the information available and up to date? If not, I created an issue at <https://github.com/JabRef/user-documentation/issues> or, even better, I submitted a pull request to the documentation repository.
-","['CHANGELOG.md', 'src/main/java/org/jabref/gui/entryeditor/EntryEditorPreferences.java', 'src/main/java/org/jabref/gui/entryeditor/FieldsEditorTab.java', 'src/main/java/org/jabref/preferences/JabRefPreferences.java']"
-repository0,11269,2024-04-30T13:25:33Z,2024-05-02T09:39:48Z,Yery,Selected group loses focus after paste doi,"JabRef 5.14--2024-04-30--4f87740
-Windows 11 10.0 amd64 
-Java 21.0.2 
-JavaFX 22.0.1+7
-
-### Checked with the latest development build (copy version output from About dialog)
-
-- [X] I made a backup of my libraries before testing the latest development version.
-- [X] I have tested the latest development version and the problem persists
-
-### Steps to reproduce the behaviour
-
-1. Select a group
-2. Copy a doi from somewhere 
-3. Click anywhere in the table and paste, jabref will correctly create a new entry with the selected group assigned
-4. Copy another doi from somewhere
-5. Click anywhere in the table and paste, jabref will create a new entry but not assigned to the group. In fact it will not show up in ""all entries"" immediately, and one needs to click a few random groups back and forth, until that works. But still, the new entry is not assigned to the group that was selected first.
-
-Note: 
-This is related to issue #8933, it seems that after the first copy-paste in both instances, the group is somehow ""deselected"", even though it is still shown as selected.
-Issue  #11085 might be related too
-","Thanks for the report, will be fixed by https://github.com/JabRef/jabref/pull/11271|SEPARATOR|Great thanks a lot! Does it fix issue #8933 too?|SEPARATOR|You can test this version https://builds.jabref.org/pull/11271/merge
-
-I will test the other issue as well but I think it was the same underlying cause |SEPARATOR|I just tried it and can confirm that it fixes all three related issues #8933, #11271 and #11085! Thanks again.|SEPARATOR|",Siedlerchr|SEPARATOR|Yery|SEPARATOR|Siedlerchr|SEPARATOR|Yery,4,"bug, groups, import",11271,2024-05-02T09:39:47Z,Siedlerchr,Fix import selected groups,"Fixes #11085 
-Fixes #11269
-Fixes #8933
-
-Equals/hashcode changes too much
-
-
-1. Create a new empty library
-2. add a subgroup
-3. Paste two or more entries in the subgroup
-4. 
-
-<!-- 
-Describe the changes you have made here: what, why, ... 
-Link the issue that will be closed, e.g., ""Closes #333"".
-If your PR closes a koppor issue, link it using its URL, e.g., ""Closes https://github.com/koppor/jabref/issues/47"".
-""Closes"" is a keyword GitHub uses to link PRs with issues; do not change it.
-Don't reference an issue in the PR title because GitHub does not support auto-linking there.
--->
-
-### Mandatory checks
-
-<!-- 
-- Go through the list below. Please don't remove any items.
-- [x] done; [ ] not done / not applicable
--->
-
-- [x] Change in `CHANGELOG.md` described in a way that is understandable for the average user (if applicable)
-- [ ] Tests created for changes (if applicable)
-- [x] Manually tested changed features in running JabRef (always required)
-- [ ] Screenshots added in PR description (for UI changes)
-- [ ] [Checked developer's documentation](https://devdocs.jabref.org/): Is the information available and up to date? If not, I outlined it in this pull request.
-- [ ] [Checked documentation](https://docs.jabref.org/): Is the information available and up to date? If not, I created an issue at <https://github.com/JabRef/user-documentation/issues> or, even better, I submitted a pull request to the documentation repository.
-","['CHANGELOG.md', 'src/main/java/org/jabref/gui/LibraryTab.java']"
